+++ src/lib/domain/calculator.ts
@@ -1,3 +1,14 @@
+// The different types of HSFP visas
+// TODO: Rename from A | B | C to more generic name
+export enum VisaType {
+  // Advanced academic research activities
+  A = 'A',
+  // Advanced specialized/technical activities
+  B = 'B',
+  // Advanced business management activities
+  C = 'C',
+}
+
 // A Criteria is a row in the HSP points table.
 export interface Criteria {
   id: string
