+++ src/lib/domain/form.ts
@@ -1,5 +1,5 @@
-import { VisaType } from '@lib/domain/visa'
 import { z } from 'zod'
+import { VisaType } from '@lib/domain/calculator'
 
 // TODO: decide how to split the sections
 export const SectionNameSchema = z.enum([
@@ -13,6 +13,11 @@ export const SectionNameSchema = z.enum([
   // 'university',
   // 'investor',
 ])
+
+export interface VisaProgress {
+  section: SectionName
+  promptIndex: number
+}
 
 export type SectionName = z.infer<typeof SectionNameSchema>
 
@@ -43,4 +48,22 @@ export interface NumberPrompt {
 export interface BooleanPrompt {
   id: string
   type: 'BOOLEAN'
+}
+
+export function nextStepOfForm(formConfig: FormConfig, progress: VisaProgress) {
+  const currentSection = formConfig.sections[progress.section]!!
+  const shouldUseNextCategory =
+    currentSection.length <= progress.promptIndex + 1
+  const promptIndex = shouldUseNextCategory ? 0 : progress.promptIndex + 1
+  const currentCategoryIndex = formConfig.order.findIndex(
+    it => it === progress.section,
+  )
+  const section = shouldUseNextCategory
+    ? formConfig.order[currentCategoryIndex + 1]
+    : progress.section
+
+  return {
+    section,
+    promptIndex,
+  } as VisaProgress
 }
