--- src/lib/domain/visa.ts
@@ -1,10 +1,0 @@
-// The different types of HSFP visas
-// TODO: Rename from A | B | C to more generic name
-export enum VisaType {
-  // Advanced academic research activities
-  A = 'A',
-  // Advanced specialized/technical activities
-  B = 'B',
-  // Advanced business management activities
-  C = 'C',
-}
